requirements.txt
celery[redis]
psycopg2-binary
python-dotenv
redis
requests
flower
